requirements.txt
# Django 6.0
Django>=6.0,<7.0

# Production Server
gunicorn==23.0.0

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Static Files
whitenoise==6.8.2

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv==1.0.1
requests==2.32.3
google-genai
python-pptx
python-docx
chromadb
fpdf

# PDF & HWP Processing (for school_violence RAG)
PyPDF2>=3.0.0
olefile>=0.46

# Auth & SSO
django-allauth==65.14.0
PyJWT==2.10.1

# Cloud Storage (Cloudinary)
cloudinary==1.40.0
django-cloudinary-storage==0.3.0

# Security
django-ratelimit==4.1.0

# Saju / Astronomy
korean-lunar-calendar==0.3.1
ephem==4.1.6
pytz>=2024.1
